atualizando repo

+++ src/main/java/dscatalog/services/CategoryService.java
@@ -13,7 +13,7 @@ import dscatalog.dto.CategoryDTO;
 import dscatalog.entities.Category;
 import dscatalog.repositories.CategoryRepository;
 import dscatalog.services.exceptions.DataBaseException;
-import dscatalog.services.exceptions.SourceNotFoundException;
+import dscatalog.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 
 @Service
@@ -22,62 +22,55 @@ public class CategoryService {
 	@Autowired
 	private CategoryRepository repository;
 
-	
-
-
 	@Transactional(readOnly = true)
-	public Page<CategoryDTO> findAll(PageRequest pageRequest ) {
+	public Page<CategoryDTO> findAll(PageRequest pageRequest) {
 		Page<Category> listCategory = repository.findAll(pageRequest);
 		return listCategory.map(x -> new CategoryDTO(x));
 	}
+
 	@Transactional(readOnly = true)
 	public CategoryDTO findById(Long id) {
-		Optional<Category> categoryOptinal = repository.findById(id);
-		Category category = categoryOptinal.orElseThrow(() -> new SourceNotFoundException("ID não encontrado"));
-		return new CategoryDTO(category );
-	}
-	
-	
-	@Transactional
-	public CategoryDTO save(CategoryDTO dto) {
-
-	    Category entity = new Category();
-	    entity.setName(dto.getName());
-
-	    entity = repository.save(entity);
-
-
-	    return new CategoryDTO(entity);
+		Optional<Category> obj = repository.findById(id);
+		Category entity = obj.orElseThrow(() -> new ResourceNotFoundException("ID não encontrado"));
+		return new CategoryDTO(entity);
 	}
 
-	
 	@Transactional
-	public CategoryDTO update(Long id,CategoryDTO dto) {
-		
+	public CategoryDTO insert(CategoryDTO dto) {
+
+		Category entity = new Category();
+		entity.setName(dto.getName());
+
+		entity = repository.save(entity);
+
+		return new CategoryDTO(entity);
+	}
+
+	@Transactional
+	public CategoryDTO update(Long id, CategoryDTO dto) {
+
 		try {
+			// ao inves de usar .findById(id) para evitar duas consultas no database
 			Category entity = repository.getReferenceById(id);
 			entity.setName(dto.getName());
 			entity = repository.save(entity);
 			return new CategoryDTO(entity);
+
 		} catch (EntityNotFoundException e) {
-			throw new SourceNotFoundException("id nao encontrado");
+			throw new ResourceNotFoundException("ID não encontrado para a atualização");
 		}
-		
-		
+
 	}
-	@Transactional
+
 	public void delete(Long id) {
-	    try {
-	         repository.findById(id).orElseThrow(() -> new SourceNotFoundException("ID não encontrado"));
-
-	    } catch (DataIntegrityViolationException e) {
-
-	        throw new DataBaseException("Integridade do banco de dados violada");
+	    if (!repository.existsById(id)) {
+	        throw new ResourceNotFoundException("ID não encontrado para deletar");
 	    }
-	    
-	    repository.deleteById(id);
+	    try {
+	        repository.deleteById(id);
+	    } catch (DataIntegrityViolationException e) {
+	        throw new DataBaseException("Não é possível deletar uma categoria associada a produtos");
+	    }
 	}
-
-	
 
 }
